Fixed In Averages

+++ src/main/java/u1pp/Main.java
@@ -20,7 +20,7 @@ public class Main
      * Should return a double that is mathematically accurate. 
      */
     public static double averageOfThree(int a , int b , int c) {
-    	return (a + b + c) / 3;
+    	return (double)(a + b + c) / 3;
     	
     }
